+++ devices/Paexo/include/Paexo.h
@@ -17,6 +17,8 @@
 
 #include <yarp/os/PeriodicThread.h>
 
+#include "paexo-cmake-config.h"
+
 namespace wearable {
     namespace devices {
         class Paexo;
@@ -93,6 +95,16 @@ public:
     SensorPtr<const sensor::ITorque3DSensor>
     getTorque3DSensor(const sensor::SensorName name) const override;
 
+#ifdef ENABLE_PAEXO_USE_iFEELDriver
+
+    SensorPtr<const sensor::IForceTorque6DSensor>
+    getForceTorque6DSensor(const sensor::SensorName name) const override;
+
+#else
+    inline SensorPtr<const sensor::IForceTorque6DSensor>
+    getForceTorque6DSensor(const sensor::SensorName /*name*/) const override;
+
+#endif
     // IMPLEMENTED ACTUATORS
     // ---------------------
 
@@ -101,9 +113,6 @@ public:
 
     // UNIMPLEMENTED SENSORS
     // ---------------------
-
-    inline SensorPtr<const sensor::IForceTorque6DSensor>
-    getForceTorque6DSensor(const sensor::SensorName /*name*/) const override;
 
     inline SensorPtr<const sensor::IVirtualLinkKinSensor>
     getVirtualLinkKinSensor(const sensor::SensorName /*name*/) const override;
@@ -152,11 +161,13 @@ public:
 
 };
 
+#ifndef ENABLE_PAEXO_USE_iFEELDriver
 inline wearable::SensorPtr<const wearable::sensor::IForceTorque6DSensor>
 wearable::devices::Paexo::getForceTorque6DSensor(const sensor::SensorName /*name*/) const
 {
     return nullptr;
 }
+#endif
 
 inline wearable::SensorPtr<const wearable::sensor::IVirtualLinkKinSensor>
 wearable::devices::Paexo::getVirtualLinkKinSensor(const sensor::SensorName /*name*/) const